pyproject.toml
[project]
name = "soc-design-ai-copilot"
version = "0.1.0"
description = "RAG + Agent assistant for SoC/RTL designers, with retrieval & faithfulness eval pipeline."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ogata" }]
dependencies = [
    "langchain>=1.0",
    "langchain-classic",
    "langchain-openai",
    "langchain-community",
    "langchain-chroma",
    "langchain-text-splitters",
    "langgraph",
    "python-dotenv",
    "pypdf",
    "rank-bm25",
    "flashrank",
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "ipykernel"]

[project.scripts]
copilot-ingest = "soc_copilot.rag.ingest:main"
copilot-eval-retrieval = "soc_copilot.eval.run_retrieval:main"
copilot-eval-faithfulness = "soc_copilot.eval.run_faithfulness:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
